Changes and Fixes
Added a few new commands such as datawipe, party, and pheal. Fixed some commands to work better. Fixed loading issue with CombatInstance.
