+++ DiscomonProject/Discord/Commands/BasicCommands.cs
@@ -12,60 +12,55 @@ namespace DiscomonProject.Discord
 {
     public class BasicCommands : ModuleBase<SocketCommandContext>
     {
-        [Command("ping")]
-        public async Task Ping()
-        {
-            await ReplyAsync("pong");
-        }
-
-        [Command("whoami")]
-        public async Task ShowAccount()
-        {
-            UserAccount acc = UserHandler.GetUser(Context.User.Id);
-
-            await ReplyAsync($"Here is your information: \nID: {acc.UserId}\nName: {acc.Name}\nAvatar: {acc.AvatarUrl}\nGuildID: {acc.Char.CurrentGuildId}");
-        }
 
         [Command("monstat")]
-        public async Task MonStat(int num)
+        public async Task MonStat([Remainder]int num)
         {
-            ContextIds ids = new ContextIds(Context);
-            UserAccount user = UserHandler.GetUser(ids.UserId);
+            ContextIds idList = new ContextIds(Context);
+            UserAccount user = UserHandler.GetUser(idList.UserId);
+
+            //Tests each case to make sure all circumstances for the execution of this command are valid (character exists, in correct location)
+            try
+            {
+                await UserHandler.CharacterExists(idList);
+                await UserHandler.ValidCharacterLocation(idList);
+            }
+            catch(InvalidCharacterStateException)
+            {
+                return;
+            }
             
             await Context.Channel.SendMessageAsync(
             "",
-            embed: MonEmbedBuilder.MonStats((BasicMon)user.Char.Party[0]))
+            embed: MonEmbedBuilder.MonStats((BasicMon)user.Char.Party[num]))
             .ConfigureAwait(false);
         }
 
-        [Command("debuginfo")]
-        public async Task DebugInfo(SocketGuildUser target)
+        [Command("party")]
+        public async Task Party()
         {
-            ContextIds ids = new ContextIds(Context);
-            UserAccount user;
+            ContextIds idList = new ContextIds(Context);
+            var user = UserHandler.GetUser(idList.UserId);
 
-            if(target != null)
+            //Tests each case to make sure all circumstances for the execution of this command are valid (character exists, in correct location)
+            try
             {
-                user = UserHandler.GetUser(target.Id);
+                await UserHandler.CharacterExists(idList);
+                await UserHandler.ValidCharacterLocation(idList);
             }
-            else
+            catch(InvalidCharacterStateException)
             {
-                user = UserHandler.GetUser(ids.UserId);
+                return;
             }
 
-            await MessageHandler.SendMessage(ids.GuildId, ids.ChannelId, user.DebugString());
-        }
+            var count = 6;
+            foreach(BasicMon mon in user.Char.Party)
+            {
+                await MessageHandler.SendEmbedMessage(idList, "", MonEmbedBuilder.FieldMon(mon));
+                count--;
+            }
 
-        [Command("debugreset")]
-        public async Task DebugReset()
-        {
-            ContextIds ids = new ContextIds(Context);
-            var user = UserHandler.GetUser(ids.UserId);
-            user.Char = null;
-            user.HasCharacter = false;
-            user.PromptState = -1;
-
-            await MessageHandler.SendMessage(ids, $"{user.Mention}, your character has been reset.");
+            await MessageHandler.SendEmbedMessage(idList, "", MonEmbedBuilder.EmptyPartySpot(count));
         }
 
         [Command("enter")]
@@ -175,143 +170,5 @@ namespace DiscomonProject.Discord
             }
         }
 
-        [Command("duel")]
-        public async Task Duel(SocketGuildUser target)
-        {
-            var fromUser = UserHandler.GetUser(Context.User.Id);
-            var toUser = UserHandler.GetUser(target.Id);
-
-            ContextIds idList = new ContextIds(Context);
-
-            //Tests each case to make sure all circumstances for the execution of this command are valid (character exists, in correct location)
-            try
-            {
-                await UserHandler.CharacterExists(idList);
-                await UserHandler.OtherCharacterExists(idList, toUser);
-                await UserHandler.ValidCharacterLocation(idList);
-                await UserHandler.OtherCharacterLocation(idList, toUser);
-            }
-            catch(InvalidCharacterStateException)
-            {
-                return;
-            }
-
-            //Check that the user did not target themself with the command
-            if(fromUser.UserId != toUser.UserId)
-            {
-                //Set the current user's combat request ID to the user specified
-                fromUser.Char.CombatRequest = toUser.UserId;
-
-                //Check if the specified user has a combat request ID that is the current user's ID
-                if(toUser.Char.CombatRequest == fromUser.UserId)
-                {
-                    //Make sure neither users are in combat while sending response request
-                    if(fromUser.Char.InCombat)
-                    {
-                        await Context.Channel.SendMessageAsync($"{Context.User.Mention}, you cannot start a duel while in combat!");
-                    }
-                    else if(toUser.Char.InCombat)
-                    {
-                        await Context.Channel.SendMessageAsync($"{Context.User.Mention}, you cannot start a duel with a player who is in combat!");
-                    }
-                    else
-                    {
-                        //Start duel
-                        CombatInstance combat = new CombatInstance(idList, Context.User.Id, target.Id);
-
-                        await Context.Channel.SendMessageAsync($"The duel between {target.Mention} and {Context.User.Mention} will now begin!");
-                        fromUser.Char.InCombat = true;
-                        fromUser.Char.InPvpCombat = true;
-                        fromUser.Char.CombatRequest = 0;
-                        fromUser.Char.InCombatWith = toUser.UserId;
-                        fromUser.Char.Combat = new CombatInstance(idList, fromUser.UserId, toUser.UserId);
-
-                        toUser.Char.InCombat = true;
-                        toUser.Char.InPvpCombat = true;
-                        toUser.Char.CombatRequest = 0;
-                        toUser.Char.InCombatWith = fromUser.UserId;
-                        toUser.Char.Combat = new CombatInstance(idList, toUser.UserId, fromUser.UserId);
-
-                        await CombatHandler.StartCombat(fromUser.Char.Combat);
-                    }
-                }
-                else
-                {
-                    //Make sure neither users are in combat while sending initial request
-                    if(fromUser.Char.InCombat)
-                    {
-                        await Context.Channel.SendMessageAsync($"{Context.User.Mention}, you cannot request a duel when you are in combat!");
-                    }
-                    else if(toUser.Char.InCombat)
-                    {
-                        await Context.Channel.SendMessageAsync($"{Context.User.Mention}, you cannot duel a player who is in combat!");
-                    }
-                    else
-                    {
-                        //Challenge the specified user
-                        await Context.Channel.SendMessageAsync($"{target.Mention}, you have been challenged to a duel by {Context.User.Mention}\nUse the \"duel [mention target]\" command to accept.");
-                    }
-                }
-            }
-            else
-            {
-                //Tell the current user they have are a dum dum
-                await Context.Channel.SendMessageAsync($"{Context.User.Mention}, you cannot duel yourself.");
-            }
-        }
-
-        [Command("attack")]
-        public async Task Attack()
-        {
-            var user = UserHandler.GetUser(Context.User.Id);
-            ContextIds idList = new ContextIds(Context);
-            
-            if(user.Char.Combat == null)
-            {
-                await MessageHandler.SendMessage(idList, $"{user.Mention} aren't in combat right now!");
-            }
-            else
-            {
-                await CombatHandler.Attack(user.Char.Combat);
-            }
-        }
-
-        [Command("exitcombat")]
-        public async Task ExitCombat()
-        {
-            var user = UserHandler.GetUser(Context.User.Id);
-            
-            ContextIds idList = new ContextIds(Context);
-            
-            //Tests each case to make sure all circumstances for the execution of this command are valid (character exists, in correct location)
-            try
-            {
-                await UserHandler.CharacterExists(idList);
-                await UserHandler.ValidCharacterLocation(idList);
-            }
-            catch(InvalidCharacterStateException)
-            {
-                return;
-            }
-
-            if(user.Char.InPvpCombat)
-            {
-                var opponent = UserHandler.GetUser(user.Char.InCombatWith);
-
-                user.Char.ExitCombat();
-                opponent.Char.ExitCombat();
-
-                await Context.Channel.SendMessageAsync($"{Context.User.Mention} has forfeited the match! {opponent.Char.Name} wins by default.");
-            }
-            else if(user.Char.InCombat)
-            {
-                await Context.Channel.SendMessageAsync($"{Context.User.Mention} blacked out!");
-            }
-            else
-            {
-                await Context.Channel.SendMessageAsync($"{Context.User.Mention}, you cannot exit combat if you are not in combat.");
-            }
-            
-        }
     }
 }
