+++ DiscomonProject/Discord/Commands/DebugCommands.cs
@@ -1,0 +1,67 @@
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Drawing;
+using System.Threading.Tasks;
+using DiscomonProject.Storage.Implementations;
+using Discord;
+using Discord.Commands;
+using Discord.WebSocket;
+
+namespace DiscomonProject.Discord
+{
+    public class DebugCommands : ModuleBase<SocketCommandContext>
+    {
+        [Command("ping")]
+        public async Task Ping()
+        {
+            await ReplyAsync("pong");
+        }
+
+        [Command("whoami")]
+        public async Task ShowAccount()
+        {
+            UserAccount acc = UserHandler.GetUser(Context.User.Id);
+
+            await ReplyAsync($"Here is your information: \nID: {acc.UserId}\nName: {acc.Name}\nAvatar: {acc.AvatarUrl}\nGuildID: {acc.Char.CurrentGuildId}");
+        }
+
+        [Command("debuginfo")]
+        public async Task DebugInfo(SocketGuildUser target)
+        {
+            ContextIds ids = new ContextIds(Context);
+            UserAccount user;
+
+            if(target != null)
+            {
+                user = UserHandler.GetUser(target.Id);
+            }
+            else
+            {
+                user = UserHandler.GetUser(ids.UserId);
+            }
+
+            await MessageHandler.SendMessage(ids.GuildId, ids.ChannelId, user.DebugString());
+        }
+
+        [Command("debugresetchar")]
+        public async Task DebugResetCharacter()
+        {
+            ContextIds ids = new ContextIds(Context);
+            var user = UserHandler.GetUser(ids.UserId);
+            user.Char = null;
+            user.HasCharacter = false;
+            user.PromptState = -1;
+
+            await MessageHandler.SendMessage(ids, $"{user.Mention}, your character has been deleted.");
+        }
+
+        [Command("datawipe")]
+        public async Task DataWipe()
+        {
+            ContextIds idList = new ContextIds(Context);
+            await MessageHandler.SendMessage(idList, "User data cleared. Reboot bot to take effect.");
+            UserHandler.ClearUserData();
+        }
+    }
+}
